Get the Zigbee Coordinator building again
A change landed in esphome 2023.4 that changed the behavior of how
ASyncTCP is pulled in. It's now included by default, so it conflicts
with the include in the stream-server.

See https://github.com/tube0013/esphome-stream-server-v2/pull/7

I switched to pulling the fork mentioned in the above PR to get this
building again. When PR#7 is accepted, I'll go back to the main
upstream.